requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
Flask==2.0.2
idna==3.3
importlib-metadata==4.8.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pkg-resources==0.0.0
pycodestyle==2.8.0
pycrypto==2.6.1
requests==2.26.0
toml==0.10.2
typing-extensions==4.0.1
urllib3==1.26.7
Werkzeug==2.0.2
zipp==3.6.0
